adds custom templates and better file creation

+++ src/MaestroRokuCommands.ts
@@ -17,10 +17,11 @@ export class MaestroRokuCommands {
   private context: vscode.ExtensionContext;
   private host: string;
   private extensionTemplates: any;
-  private workspaceTemplates: any;
+  private resourcesPath: string;
 
   registerCommands(context: vscode.ExtensionContext) {
     this.context = context;
+    this.resourcesPath = context.asAbsolutePath(path.join('resources'));
     this.extensionTemplates = this.loadExtensionTemplatesJson(context.asAbsolutePath(path.join('resources', 'templates')));
     //TODO get these from local workspace
     // this.workspaceTemplates = this.loadExtensionTemplatesJson(context.asAbsolutePath(path.join('resources','templates')));
@@ -76,6 +77,9 @@ export class MaestroRokuCommands {
     subscriptions.push(vscode.commands.registerCommand('maestro.files.createFile', (selectedFile) => {
       this.createFile(selectedFile);
     }));
+    subscriptions.push(vscode.commands.registerCommand('maestro.files.createCustomTemplates', (selectedFile) => {
+      this.createCustomTemplates();
+    }));
   }
   async createFile(selectedFile) {
     if (selectedFile) {
@@ -103,12 +107,26 @@ export class MaestroRokuCommands {
     }
   }
 
+
+  private async getTemplates(): Promise<any> {
+    let localTemplatePath = await this.fileUtils.getLocalTemplatesPath();
+    let templates = this.extensionTemplates;
+    if (fs.existsSync(localTemplatePath)) {
+      try {
+        templates = this.loadExtensionTemplatesJson(localTemplatePath);
+      } catch (error) {
+        console.error('Unable to load custom templates', error);
+      }
+    }
+    return templates;
+  }
   async createComponent(selectedFile) {
     if (!selectedFile) {
       vscode.window.showErrorMessage('Select a path in the explorer');
     }
     let targetPath = this.getTargetPathForFileCreation(selectedFile);
-    let items = this.extensionTemplates.templates.filter((template) => template.files.length > 0).map((item) => {
+    let sourceRootForSelectedFile = await this.getSourceRootForSelectedPath(selectedFile);
+    let items = (await this.getTemplates()).templates.filter((template) => template.files.length > 0).map((item) => {
       return {
         label: item.name,
         item: item
@@ -143,28 +161,44 @@ export class MaestroRokuCommands {
 
       let sourcePkgPath = this.fileUtils.getPkgPathFromFilePath(path.join(targetPath, `${className}`));
 
-      const targetFilePath = await util.showQuickPickInputBox({
-        placeholder: 'Confirm path name',
-        items: [{ label: this.fileUtils.getPkgPathFromFilePath(targetPath) }]
+      let editedTargetPath = await vscode.window.showInputBox({
+        title: 'Path where new files are created',
+        placeHolder: targetPath,
+        valueSelection: [targetPath.length - 1, targetPath.length - 1],
+        value: this.fileUtils.getPkgPathFromFilePath(targetPath)
       });
 
-      //TODO if path does not exist, then create it
+      if (!editedTargetPath) {
+        return;
+      }
+
+      targetPath = path.join(sourceRootForSelectedFile, editedTargetPath);
+      this.fileUtils.ensurePathExists(targetPath);
 
       for (let filePath of selectedItem.files) {
-        let targetFilePath = this.getTargetFilePath(filePath, targetPath, className);
-        if (!rootFilePath) {
-          rootFilePath = targetFilePath;
-        }
-        console.log('creating file', targetFilePath);
-        if (fs.existsSync(targetFilePath)) {
-          vscode.window.showWarningMessage(`${targetFilePath} already exists - not generating`);
-          continue;
-        }
-        let text = this.loadAndReplaceExtensionText(filePath, namespaceName, className, sourcePkgPath);
-        //the fileExtension is everything from ., after the last path.sep in filePath
+        try {
 
-        fs.writeFileSync(targetFilePath, text);
-        vscode.window.showInformationMessage(`Created ${targetFilePath}`);
+          let targetFilePath = this.getTargetFilePath(filePath, targetPath, className);
+          if (!rootFilePath) {
+            rootFilePath = targetFilePath;
+          }
+          console.log('creating file', targetFilePath);
+          if (fs.existsSync(targetFilePath)) {
+            vscode.window.showWarningMessage(`${targetFilePath} already exists - not generating`);
+            continue;
+          }
+          if (fs.existsSync(filePath)) {
+            vscode.window.showWarningMessage(`Template file: ${filePath} does not exist. Please check your custom templates.json`);
+            continue;
+          }
+          let text = this.loadAndReplaceExtensionText(filePath, namespaceName, className, sourcePkgPath);
+          //the fileExtension is everything from ., after the last path.sep in filePath
+
+          fs.writeFileSync(targetFilePath, text);
+          vscode.window.showInformationMessage(`Created ${targetFilePath}`);
+        } catch (error) {
+          vscode.window.showErrorMessage(`Could not create template file`);
+        }
       }
     }
     if (rootFilePath) {
@@ -175,6 +209,32 @@ export class MaestroRokuCommands {
 
     }
   }
+
+  private async createCustomTemplates() {
+    let localTemplatePath = await this.fileUtils.getLocalTemplatesPath();
+
+    if (!localTemplatePath) {
+      return;
+    }
+    if (fs.existsSync(localTemplatePath)) {
+      vscode.window.showInformationMessage(`Templates folder already exists`);
+      return;
+    }
+
+    let extensionTemplatesPath = path.join(this.resourcesPath, 'templates');
+    //create localTemplatePath
+    fs.mkdirSync(localTemplatePath, { recursive: true });
+
+    //copy all files from extensionTemplatesPath to localTemplatePath
+
+    let templateFiles = fs.readdirSync(extensionTemplatesPath);
+    for (let file of templateFiles) {
+      fs.copyFileSync(path.join(extensionTemplatesPath, file), path.join(localTemplatePath, file));
+    }
+
+    vscode.window.showInformationMessage(`Copied Maestro template files to your .vscode folder for customization`);
+  }
+
   getTargetPathForFileCreation(selectedFile: any) {
     //TODO - what if there is no selectedFile
     let parsedTargetPath = path.parse(selectedFile.fsPath);
@@ -182,13 +242,17 @@ export class MaestroRokuCommands {
     return targetPath;
   }
 
+  async getSourceRootForSelectedPath(selectedFile: any) {
+    let pkgPath = await this.fileUtils.getPkgPathFromFilePath(selectedFile);
+    return selectedFile.fsPath.replace(pkgPath, '');
+  }
 
   private getTargetFilePath(filePath: string, targetPath: string, className: string) {
     let templateFileName = filePath.split(path.sep).slice(-1)[0].split('.');
     templateFileName.shift();
     let templateExtension = templateFileName.join('.');
-        let targetFilename = `${className}.${templateExtension}`;
-        return path.join(targetPath, targetFilename);
+    let targetFilename = `${className}.${templateExtension}`;
+    return path.join(targetPath, targetFilename);
   }
 
   private loadAndReplaceExtensionText(filePath: string, namespaceName: string, className: string, sourcePkgPath: string) {
@@ -210,7 +274,7 @@ export class MaestroRokuCommands {
     return json;
   }
 
-    async onCycleFiles() {
+  async onCycleFiles() {
     if (vscode.window.activeTextEditor) {
       const currentDocument = vscode.window.activeTextEditor.document;
       let targetFilename = this.fileUtils.getAlternateFileName(currentDocument.fileName);
